monthly analysis added
